Re-factor the python scripts to support Google Cloud Run structure
