+++ functions/show_availability_api/requirements.txt
@@ -1,0 +1,4 @@
+functions-framework==3.8.1
+gcsfs==2024.10.0
+polars==1.11.0
+requests==2.32.3
