+++ requirements.txt
@@ -1,4 +1,8 @@
+aiohappyeyeballs==2.4.3
+aiohttp==3.10.10
+aiosignal==1.3.1
 asttokens==2.4.1
+attrs==24.2.0
 blinker==1.8.2
 cachetools==5.5.0
 certifi==2024.8.30
@@ -12,9 +16,13 @@ decorator==5.1.1
 deprecation==2.1.0
 executing==2.1.0
 Flask==3.0.3
+frozenlist==1.5.0
+fsspec==2024.10.0
 functions-framework==3.8.1
+gcsfs==2024.10.0
 google-api-core==2.21.0
 google-auth==2.35.0
+google-auth-oauthlib==1.2.1
 google-cloud-core==2.4.1
 google-cloud-storage==2.18.2
 google-crc32c==1.6.0
@@ -30,12 +38,15 @@ jupyter_client==8.6.3
 jupyter_core==5.7.2
 MarkupSafe==3.0.2
 matplotlib-inline==0.1.7
+multidict==6.1.0
 nest-asyncio==1.6.0
+oauthlib==3.2.2
 packaging==24.1
 parso==0.8.4
 platformdirs==4.3.6
 polars==1.11.0
 prompt_toolkit==3.0.48
+propcache==0.2.0
 proto-plus==1.24.0
 protobuf==5.28.2
 psutil==6.0.0
@@ -47,6 +58,7 @@ python-dateutil==2.9.0.post0
 pywin32==308
 pyzmq==26.2.0
 requests==2.32.3
+requests-oauthlib==2.0.0
 rsa==4.9
 six==1.16.0
 stack-data==0.6.3
@@ -57,3 +69,4 @@ urllib3==2.2.3
 watchdog==5.0.3
 wcwidth==0.2.13
 Werkzeug==3.0.4
+yarl==1.17.1
